+++ firmware/lib/INA236/INA236.h
@@ -22,7 +22,7 @@ class INA236 : public Sensor
     void init(bool adcRange, uint16_t sampleAverage, uint16_t shuntResistanceVal);
 
     void readShuntVoltage(double &shuntVoltage);
-    void readBusVoltage(uint32_t &busVoltage);
+    void readBusVoltage(double &busVoltage);
     void readPower(uint32_t &power);
     void readCurrent(double &current);
